Move meta.DMID to its own file
